perf(find-wallet): poc detail-route modals + defer desktop filter hydration
