+++ app/[locale]/wallets/find-wallet/@modal/default.tsx
@@ -1,0 +1,4 @@
+// No modal rendered unless the intercepting route matches.
+const Default = () => null
+
+export default Default
